test: 예약 시간 서비스 테스트 수정
